# gh-103193: Speedup and inline `inspect._is_type`